# apps/bot-server/src/api/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"telegramId": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"firstName": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					},
					"lastMeasurementAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"chats": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"telegramId",
					"username",
					"firstName",
					"value",
					"lastMeasurementAt",
					"chats",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UsersListResponse": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/UserDto"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"total",
					"page",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserRequest": {
				"properties": {
					"value": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string",
						"nullable": true
					},
					"firstName": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ConditionDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"isEnabled": {
						"type": "boolean"
					},
					"chance": {
						"type": "number",
						"format": "double"
					},
					"minDelta": {
						"type": "number",
						"format": "double"
					},
					"maxDelta": {
						"type": "number",
						"format": "double"
					},
					"config": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"isProtected": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"code",
					"name",
					"description",
					"isEnabled",
					"chance",
					"minDelta",
					"maxDelta",
					"config",
					"isProtected",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateConditionRequest": {
				"properties": {
					"code": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"isEnabled": {
						"type": "boolean"
					},
					"chance": {
						"type": "number",
						"format": "double"
					},
					"minDelta": {
						"type": "number",
						"format": "double"
					},
					"maxDelta": {
						"type": "number",
						"format": "double"
					},
					"config": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					}
				},
				"required": [
					"code",
					"name",
					"chance",
					"minDelta",
					"maxDelta"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateConditionRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"isEnabled": {
						"type": "boolean"
					},
					"chance": {
						"type": "number",
						"format": "double"
					},
					"minDelta": {
						"type": "number",
						"format": "double"
					},
					"maxDelta": {
						"type": "number",
						"format": "double"
					},
					"config": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "bot-server",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "ListUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsersListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortDir",
						"required": false,
						"schema": {
							"default": "desc",
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			}
		},
		"/users/{id}": {
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserRequest"
							}
						}
					}
				}
			}
		},
		"/conditions": {
			"get": {
				"operationId": "ListConditions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ConditionDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conditions"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCondition",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConditionDto"
								}
							}
						}
					}
				},
				"tags": [
					"Conditions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateConditionRequest"
							}
						}
					}
				}
			}
		},
		"/conditions/available-codes": {
			"get": {
				"operationId": "ListAvailableCodes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Conditions"
				],
				"security": [],
				"parameters": []
			}
		},
		"/conditions/{id}": {
			"patch": {
				"operationId": "UpdateCondition",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConditionDto"
								}
							}
						}
					}
				},
				"tags": [
					"Conditions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateConditionRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCondition",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Conditions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}